test: fix test cases
There appears to be invalid test cases included in the spec at:
https://github.com/bitcoincashorg/bitcoincash.org/blob/master/spec/cashaddr.md#larger-test-vectors

Specifically the first byte of the payload is the version byte:
- the most signifigant bit is reserved
- the next four bits is for the type (2 ^ 4) = 16 types
- the last three bits is for the size (2 ^ 3) = 8 sizes

As the type and the size are both the same, the first byte of the
payload should be the same, however this isn't the case. It's not clear
if these are intended to be invalid test cases.

These test cases are also not included at:
https://github.com/Bitcoin-ABC/bitcoin-abc/blob/master/src/test/cashaddrenc_tests.cpp#L310

So it looks currently best to not include these in the valid test suite.

Other errors were from typos of the version number, and have been corrected.