test: add test cases w/ number in prefix
Numbers were not in prefix as added in this commit:
 https://github.com/Bitcoin-ABC/bitcoin-abc/commit/5092e68b6dfceebc701e5659f3cedd42dcccbf41

However the specification at:
https://github.com/bitcoincashorg/bitcoincash.org/blob/master/spec/cashaddr.md
does not mention this, and there isn't clear rationale for the limitation.

The specification for bech32 specifically mentions that all US-ASCII characters are
available in the human readable portion (hrp), however there isn't mention
in cashaddr spec. Nor is there mention for the max length of the prefix.