test: add test vectors for cashaddr
Test vectors come from:
- https://github.com/bitcoincashorg/bitcoincash.org/blob/d10967696bfc2ea74d7dac71b4f5e51410a9c00d/spec/cashaddr.md
- https://github.com/Steve132/bitcoincash.org/blob/f07642c78acebe80d40416928f91de95393c1dbb/spec/cashaddr.md
- https://github.com/bcoin-org/bcash/blob/83e61553a66a04b1f3c3b967bbe5db152de0999b/test/address-test.js
- https://github.com/Bitcoin-ABC/bitcoin-abc/blob/a67d106b8c7f64c7555aa3f7a2eb118b30c50503/src/test/cashaddrenc_tests.cpp

With additional vectors added for valid and invalid encoding and
decoding. This is to cover cases such as uppercase, leading zero
hashes, null buffers, missing and short prefix, maximum lengths
for prefixes and data buffers. Also invalid types, prefix, and
hash data with error handling.

A note should be made about numbers in the prefix, and that
they are not compatible in the prefix. This was added in commit:
https://github.com/Bitcoin-ABC/bitcoin-abc/commit/5092e68b6dfceebc701e5659f3cedd42dcccbf41
However has not been reflected in the specification at:
https://github.com/bitcoincashorg/bitcoincash.org/blob/d10967696bfc2ea74d7dac71b4f5e51410a9c00d/spec/cashaddr.md

The specification for bech32 specifically mentions that all US-ASCII
characters are available in the human readable portion (hrp) with a
maximum length of 83. In the case for cashaddr this is the prefix.
The specification for cashaddr does not yet specify these requirements.

The requirement for the prefix not including a number is likely a
result of numbers in the prefix generating invalid checksums, as
can be found by the test case pr3f1x:prgq00r9x8r3q09aj0g75wmjjr7tkvreusdmpmxgtw
as defined in the invalid decoding vectors.

The requirement for the length of the prefix is also not defined
in the cashaddr spec so a value of 83 has been used here to match
the bech32 spec instead.